+++ packages/dev-server/e2e/custom-vite-base/vite.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vite'
+import devServer from '../../src'
+
+export default defineConfig(async () => {
+  return {
+    base: '/docs/',
+    plugins: [
+      devServer({
+        entry: '../basic/mock/worker.ts',
+      }),
+    ],
+  }
+})