+++ packages/dev-server/playwright.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig, devices } from '@playwright/test'
-
 export default defineConfig({
+  testDir: './e2e',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
   workers: process.env.CI ? 1 : undefined,
   use: {
-    baseURL: 'http://localhost:6173',
+    trace: 'on-first-retry',
   },
   projects: [
     {
@@ -16,9 +16,4 @@ export default defineConfig({
       retries: 2,
     },
   ],
-  webServer: {
-    command: 'yarn vite --port 6173 -c ./vite.config.ts',
-    port: 6173,
-    reuseExistingServer: !process.env.CI,
-  },
 })