+++ packages/ssg/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @hono/vite-ssg
 
+## 0.3.3
+
+### Patch Changes
+
+- [#330](https://github.com/honojs/vite-plugins/pull/330) [`11107bc013ac40c891b4a57d0ef718d822b1fa2c`](https://github.com/honojs/vite-plugins/commit/11107bc013ac40c891b4a57d0ef718d822b1fa2c) Thanks [@ryuapp](https://github.com/ryuapp)! - Add vite-plugin keyword for [Vite plugin registry](https://registry.vite.dev/)
+
+## 0.3.2
+
+### Patch Changes
+
+- [#325](https://github.com/honojs/vite-plugins/pull/325) [`3cffe86130d30909ad7799cc8adb157c47a0ae3a`](https://github.com/honojs/vite-plugins/commit/3cffe86130d30909ad7799cc8adb157c47a0ae3a) Thanks [@3w36zj6](https://github.com/3w36zj6)! - include Node.js builtin modules explicitly
+
+## 0.3.1
+
+### Patch Changes
+
+- [#318](https://github.com/honojs/vite-plugins/pull/318) [`9bc8f280de0922aa461e715d3751294425313c80`](https://github.com/honojs/vite-plugins/commit/9bc8f280de0922aa461e715d3751294425313c80) Thanks [@3w36zj6](https://github.com/3w36zj6)! - respect resolve options when creating server
+
+## 0.3.0
+
+### Minor Changes
+
+- [#304](https://github.com/honojs/vite-plugins/pull/304) [`1a6b277b6ecc58450113b63f6b3bf2e2e4fce35a`](https://github.com/honojs/vite-plugins/commit/1a6b277b6ecc58450113b63f6b3bf2e2e4fce35a) Thanks [@epelz](https://github.com/epelz)! - add ssg build option to specify extensionMap
+
+## 0.2.1
+
+### Patch Changes
+
+- [#299](https://github.com/honojs/vite-plugins/pull/299) [`5ff83b02ab87903af8e3d656df7a37dc4a43384b`](https://github.com/honojs/vite-plugins/commit/5ff83b02ab87903af8e3d656df7a37dc4a43384b) Thanks [@yusukebe](https://github.com/yusukebe)! - fix: set `enforce` as `post`
+
+## 0.2.0
+
+### Minor Changes
+
+- [#283](https://github.com/honojs/vite-plugins/pull/283) [`4c28821fbf889deba5e10c7fd7f81f50530431c7`](https://github.com/honojs/vite-plugins/commit/4c28821fbf889deba5e10c7fd7f81f50530431c7) Thanks [@3w36zj6](https://github.com/3w36zj6)! - feat: add support for ssg plugins
+
 ## 0.1.2
 
 ### Patch Changes